refactor(EnhancedClient-Tests): Move Data Models
